CHROMIUM: hardlockup: detect hard lockups without NMIs using secondary cpus
Emulate NMIs on systems where they are not available by using timer
interrupts on other cpus.  Each cpu will use its softlockup hrtimer
to update a per-cpu timestamp and then check the timestamp on the
next cpu.

The emulation uses the softlockup threshold instead of the hardlockup
threshold because the implementation runs at the same resolution as
the softlockup detector.

This patch is useful on systems where the hardlockup detector is not
available due to a lack of NMIs, for example most ARM SoCs.
Without this patch any cpu stuck with interrupts disabled can
cause a hardware watchdog reset with no debugging information,
but with this patch the kernel can detect the lockup and panic,
which can result in useful debugging info.

Signed-off-by: Colin Cross <ccross@android.com>
(cherry picked from android-exynos-manta-3.4-jb-mr2 commit
	a4b412105410659bc134ff003ba98a99a9fff698)

BUG=chrome-os-partner:22017
TEST=Boot on Snow/Peach. Run any test case and if there is a
hard lockup, system should throw a panic and dump the PC of all
CPUs.

[prathyush: rebased to chromeos-3.8 kernel with minor changes]
Change-Id: I0e8eef72691857e0c3c078548f9dd946042a7249
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/66195
Tested-by: <prathyush@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>