CHROMIUM: devfreq: Round frequencies in transition stats
Transition statistics required that the new frequency was *exactly*
as it was in the table. However, we may unexpectedly select a
frequency that is not in the table (i.e. thermal limit forces us to
select a bit lower frequency, the original DVFS table does not hold
for the current PLL for some reason, etc.)

This patch modifies transition table to select the best candidate
from the frequency table.

BUG=none
TEST=emerged kernel for nyan(Tegra124 Norrin) and made sure kernel boots.

Change-Id: Iab573eb785f08f28e4717b7937fdcc9e7118bb08
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/180247
Reviewed-by: Dylan Reid <dgreid@chromium.org>