UPSTREAM: kfifo API type safety
This patch enhances the type safety for the kfifo API.  It is now safe
to put const data into a non const FIFO and the API will now generate a
compiler warning when reading from the fifo where the destination
address is pointing to a const variable.

As a side effect the kfifo_put() does now expect the value of an element
instead a pointer to the element.  This was suggested Russell King.  It
make the handling of the kfifo_put easier since there is no need to
create a helper variable for getting the address of a pointer or to pass
integers of different sizes.

IMHO the API break is okay, since there are currently only six users of
kfifo_put().

The code is also cleaner by kicking out the "if (0)" expressions.

[akpm@linux-foundation.org: coding-style fixes]
Signed-off-by: Stefani Seibold <stefani@seibold.net>
Cc: Russell King <rmk@arm.linux.org.uk>
Cc: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

(cherry picked from commit 498d319bb512992ef0784c278fa03679f2f5649d)
[benzh: 3.14 rebase. Picked to fixup. Note that in upstream,
rt2800pci_txstatus_interrupt() has been renamed to
rt2800mmio_txstatus_interrupt(), and moved to
drivers/net/wireless/rt2x00/rt2800mmio.c
]
Signed-off-by: Ben Zhang <benzh@chromium.org>

Change-Id: I66518ca553c9763db9196062abc99b04d0f8a9ca